migrate HumanByte/Num to utils pkg
