+++ cmd/cmd.go
@@ -16,6 +16,7 @@ import (
 	"github.com/alwqx/sec/cmd/valuation"
 	"github.com/alwqx/sec/provider/sina"
 	"github.com/alwqx/sec/types"
+	"github.com/alwqx/sec/utils"
 	"github.com/alwqx/sec/version"
 	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
@@ -129,7 +130,7 @@ func InfoHandler(cmd *cobra.Command, args []string) error {
 	fmt.Fprintf(cmd.OutOrStdout(), "证券代码\t%s\n简称历史\t%s\n公司名称\t%s\n上市日期\t%s\n发行价格\t%.2f\n行业分类\t%s\n主营业务\t%s\n办公地址\t%s\n公司网址\t%s\n当前价格\t%.2f\n市净率PB\t%.2f\n市盈率TTM\t%.2f\n总市值  \t%s\n流通市值\t%s\n",
 		sec.ExCode, profile.HistoryName, profile.Name, profile.ListingDate, profile.ListingPrice,
 		profile.Category, profile.MainBusiness, profile.BusinessAddress, profile.WebSite,
-		profile.Current, profile.PB, profile.PeTTM, types.HumanNum(profile.MarketCap), types.HumanNum(profile.TradedMarketCap))
+		profile.Current, profile.PB, profile.PeTTM, utils.HumanNum(profile.MarketCap), utils.HumanNum(profile.TradedMarketCap))
 
 	if opts.Dividend {
 		dids, err := sina.QueryDividends(cmd.Context(), opts.Code)
