+++ cmd/quote/quote_history.go
@@ -134,8 +134,8 @@ func printQuoteHistory(quotes []*eastmoney.Quote) {
 			strconv.FormatFloat(quote.Open, 'g', -1, 64),
 			strconv.FormatFloat(quote.High, 'g', -1, 64),
 			strconv.FormatFloat(quote.Low, 'g', -1, 64),
-			types.HumanNum(float64(quote.TurnOver)),
-			types.HumanNum(float64(quote.Volume)),
+			utils.HumanNum(float64(quote.TurnOver)),
+			utils.HumanNum(float64(quote.Volume)),
 			strconv.FormatFloat(quote.Amplitude, 'g', -1, 64),
 			strconv.FormatFloat(quote.Velocity, 'g', -1, 64),
 			quote.Market.String() + quote.Code,
