+++ utils/utils.go
@@ -142,3 +142,29 @@ func JSONify(data interface{}) string {
 	}
 	return string(v)
 }
+
+func HumanByte(cap float64) (res string) {
+	if cap <= 0.0 {
+		res = " - "
+	} else if cap >= 1_073_741_824.0 { // 1024 * 1024 * 1024
+		res = fmt.Sprintf("%-.2f GB", cap/1_073_741_824.0)
+	} else if cap >= 1_048_576.0 { // 1024 * 1024
+		res = fmt.Sprintf("%-.2f MB", cap/1_048_576.0)
+	} else if cap >= 1_024.0 {
+		res = fmt.Sprintf("%-.2f KB", cap/1_024.0)
+	} else {
+		res = fmt.Sprintf("%-.0f B", cap)
+	}
+	return
+}
+
+func HumanNum(cap float64) (res string) {
+	if cap <= 0.0 {
+		res = " - "
+	} else if cap > 100_000_000.0 {
+		res = fmt.Sprintf("%-.2f亿", cap/100_000_000.0)
+	} else {
+		res = fmt.Sprintf("%-.2f万", cap/10_000.0)
+	}
+	return
+}
