+++ utils/utils_test.go
@@ -143,3 +143,23 @@ func TestSecDir(t *testing.T) {
 	require.Nil(t, err)
 	require.True(t, info.IsDir())
 }
+
+func TestHumanByte(t *testing.T) {
+	require.EqualValues(t, " - ", HumanByte(-1))
+	require.EqualValues(t, " - ", HumanByte(0.0))
+	require.EqualValues(t, "500 B", HumanByte(500))
+	require.EqualValues(t, "1.00 KB", HumanByte(1024))
+	require.EqualValues(t, "1.50 KB", HumanByte(1536))
+	require.EqualValues(t, "1.00 MB", HumanByte(1048576))
+	require.EqualValues(t, "2.50 MB", HumanByte(2621440))
+	require.EqualValues(t, "1.00 GB", HumanByte(1073741824))
+	require.EqualValues(t, "3.75 GB", HumanByte(4026531840))
+}
+
+func TestHumanNum(t *testing.T) {
+	require.EqualValues(t, " - ", HumanNum(-1))
+	require.EqualValues(t, " - ", HumanNum(0.0))
+	require.EqualValues(t, "1.00万", HumanNum(10000))
+	require.EqualValues(t, "10.09万", HumanNum(100900))
+	require.EqualValues(t, "1000.09亿", HumanNum(100009000009))
+}
